Updated branch closing script with code-review edits.
